+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidc/DummyCredentialDataProvider.kt
@@ -62,12 +62,14 @@ class DummyCredentialDataProvider(
                     claims = claims,
                     expiration = expiration,
                     scheme = credentialScheme,
+                    subjectPublicKey = subjectPublicKey,
                 )
 
                 ConstantIndex.CredentialRepresentation.PLAIN_JWT -> CredentialToBeIssued.VcJwt(
                     subject = AtomicAttribute2023(subjectId, CLAIM_GIVEN_NAME, "Susanne"),
                     expiration = expiration,
                     scheme = credentialScheme,
+                    subjectPublicKey = subjectPublicKey,
                 )
 
                 ConstantIndex.CredentialRepresentation.ISO_MDOC -> CredentialToBeIssued.Iso(
@@ -76,6 +78,7 @@ class DummyCredentialDataProvider(
                     },
                     expiration = expiration,
                     scheme = credentialScheme,
+                    subjectPublicKey = subjectPublicKey,
                 )
             }
         } else if (credentialScheme == MobileDrivingLicenceScheme) {
@@ -117,6 +120,7 @@ class DummyCredentialDataProvider(
                 issuerSignedItems = issuerSignedItems,
                 expiration = expiration,
                 scheme = credentialScheme,
+                subjectPublicKey = subjectPublicKey,
             )
         } else if (credentialScheme == EuPidScheme) {
             val subjectId = subjectPublicKey.didEncoded
@@ -142,6 +146,7 @@ class DummyCredentialDataProvider(
                         claims = claims,
                         expiration = expiration,
                         scheme = credentialScheme,
+                        subjectPublicKey = subjectPublicKey,
                     )
 
                 ConstantIndex.CredentialRepresentation.PLAIN_JWT -> CredentialToBeIssued.VcJwt(
@@ -158,6 +163,7 @@ class DummyCredentialDataProvider(
                     ),
                     expiration = expiration,
                     scheme = credentialScheme,
+                    subjectPublicKey = subjectPublicKey,
                 )
 
                 ConstantIndex.CredentialRepresentation.ISO_MDOC -> CredentialToBeIssued.Iso(
@@ -166,6 +172,7 @@ class DummyCredentialDataProvider(
                     },
                     expiration = expiration,
                     scheme = credentialScheme,
+                    subjectPublicKey = subjectPublicKey,
                 )
             }
         } else {